+++ NEWS.md
@@ -1,6 +1,7 @@
 # statgenGWAS 1.0.5.1
 
 * The plot function now has an option title.
+* The codeMarkers function now has an option MAC equivalent to the same option in runSingleTraitGwas.
 * Bug when using tibbles instead of data.frames as input for covariates is fixed.
 * The kinship function now has an option for returning a named identity matrix as output.
 * An extra check for the presence of . in a character marker matrix is added.